# drip-manager/swagger.yaml
swagger: "2.0"
info:
  description: "The Dynamic Real-time infrastructure planner (DRIP) allows application developers to seamlessly plan a customized virtual infrastructure based on application level constraints on QoS and resource budgets, provisioning the virtual infrastructure, deploy application components onto the virtual infrastructure, and start execution on demand using TOSCA."
  version: "3.0.0"
  title: "DRIP"
  contact:
    email: "z.zhao@uva.nl"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/drip-api"
schemes:
- "https"
- "http"
paths:
  /tosca_template:
    post:
      summary: "upload a tosca template description file"
      description: ""
      operationId: "uploadToscaTemplate"
      consumes:
      - "multipart/form-data"
      parameters:
      - in: "formData"
        name: "file"
        type: file
        description: "tosca Template description"
        required: true
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        405:
          description: "Invalid input"
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"
  /tosca_template/ids:
    get:
      summary: "Get all topolog template IDs"
      description: "Returns all IDss "
      operationId: "getToscaTemplateIDs"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"        
  /tosca_template/{id}:
    get:
      summary: "Find topolog template by ID"
      description: "Returns a single topolog template"
      operationId: "getToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to return"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"
    put:
      summary: "Updates exisintg topolog template"
      description: ""
      operationId: "updateToscaTemplateByID"
      consumes:
      - "multipart/form-data"
      produces:
      - "text/plain"
      parameters:
      - in: "formData"
        name: "file"
        type: file
        description: "tosca Template description"
      - name: "id"
        in: "path"
        description: "ID of topolog template to return"
        required: true
        type: "string"       
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"
    delete:
      summary: "Deletes a tosca topolog template"
      description: ""
      operationId: "deleteToscaTemplateByID"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to return"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"
  /planner/plan/{id}:
    get:
      summary: "plan tosca template"
      description: "Returns the ID of the planed topolog template"
      operationId: "planToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to plan"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"    
  /planner/ids:
    get:
      summary: "Get all topolog template IDs"
      description: "Returns all IDss "
      operationId: "getPlanToscaTemplateIDs"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"          
  /planner/{id}:
    get:
      summary: "get the plan tosca template"
      description: "Returns the plan topolog template"
      operationId: "getPlanToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template plan"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"        
  /provisioner/provision/{id}:
    get:
      summary: "provision tosca template"
      description: "provosions the operationId: Plan Tosca Template Returns the provision ID"
      operationId: "provisionPlanToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to plan"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"     
  /provisioner/{id}:
    get:
      summary: "the provisioned tosca template"
      description: "Returns the provisioned tosca template"
      operationId: "getProvisionToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to plan"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"  
  /provisioner/ids:
    get:
      summary: "Get all topolog template IDs"
      description: "Returns all IDss "
      operationId: "getProvisionedToscaTemplateIDs"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"         
...
securityDefinitions:
  drip_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost/oauth/dialog"
    flow: "implicit"
    scopes:
      write:ToscaTemplate: "modify topolog template in your account"
      read:ToscaTemplate: "read your topolog template"
      admin:User: "Grants access to admin operations"
      write:Credentials: "modify cloud credentials in your account"
      read:Credentials: "read your cloud credentials"
...